Gemini 2.0 pro
We are upgrading this nuxt 2 components module into nuxt 3 components module.

We have already migrated some components and configurations to nuxt 3
We need to proceed with the migration, step by step.

From the "components" folder, only migrate the ListMedia.vue component and the corresponding test. Also migrate any component that is required by this component or any of its dependencies.

Provide any code comments in english.